tests/submission/Infraware/Offline_Application_Cache/api_swapcache_error.html

<!DOCTYPE HTML>
<html manifest="manifest/clock.manifest">
 <head>
  <title>Offline Application Cache</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>API_swapCache_error</h3>
	</p>
	
	<hr>

	<div id="log"></div>
		 
	<script>

		if (('applicationCache' in window) && window.applicationCache !== null){
			var cache = window.applicationCache;

			test(function()
			{
				assert_throws("INVALID_STATE_ERR", function(){cache.swapCache();});
			}, "INVALID_STATE_ERR error test")
		}

	</script>

 </body>
</html>
